code cleanup;
