+++ QuizBot/QuizBotCore/Buttons & CallBacks/UserCommands.cs
@@ -5,6 +5,6 @@ namespace QuizBotCore
         public const string Help = "/help";
         public const string Feedback = "/feedback";
         public const string Level = "/level";
-        public const string ReportTask = "/reportThisTask";
+        public const string ReportTask = "/СообщитьОбОшибке";
     }
 }
